validacion Campos

+++ database/migrations/2024_02_29_095152_create_contratos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
                 $table->string('contrato_id');
                 $table->string('ubicacion_sala');
                 $table->integer('anios_contrato');
-                $table->integer('monto_contrato'); 
+                $table->decimal('monto_contrato', 10, 2); 
                 $table->boolean('bono_hospedaje_qori_loyalty')->default(false);
                 $table->boolean('bono_hospedaje_internacional')->default(false);
                 $table->decimal('valor_total_credito_directo', 10, 2)->nullable();
